feat(iOSFilters): Add searchbar and filter button, also create a component for company card
