+++ ios/GreenCircle/GreenCircle/Presentation/Views/Components/FilterView.swift
@@ -1,0 +1,18 @@
+//
+//  FilterView.swift
+//  GreenCircle
+//
+//  Created by Dan FuPo on 10/10/23.
+//
+
+import SwiftUI
+
+struct FilterView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    FilterView()
+}
